feat: add select_where to base database session manager class
